requirements.txt
Django==2.2.7
Python==3.8.0
djangorestframework==3.11.0